+++ src/components/Payment.js
