+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as Header } from "./Header";
 export { default as Footer } from "./Footer";
-export { default as Paymentinfo } from "./Paymentinfo";
-export { default as Paymentmask } from "./Paymentmask";
-export { default as Paymentrecieved } from "./Paymentrecieved";
+export { default as Payment } from "./Payment";
+export { default as LandingPage } from "./LandingPage";
+export { default as Success } from "./Success";
